requirements.txt
PyQt5
Pillow
psutil